fixed bugs, add generic type, refactore code
