+++ src/constants/translations/ua/find-offer-page.json
@@ -11,5 +11,12 @@
   "contentSwitcher": {
     "switcher-tutor": "Пропозиції від репетиторів на заняття з різних категорій та предметів",
     "switcher-student": "Пропозиції від студентів на заняття з різних категорій та предметів"
+  },
+  "radioFilter": {
+    "rating": "Рейтинг",
+    "any": "Будь-який рейтинг",
+    "5stars": "5 зірок",
+    "4stars": "4 і нижче",
+    "3stars": "3 і нижче"
   }
 }
