+++ src/types/components/radioButtonInputs/RadioButtonInputs.types.ts
@@ -1,0 +1,4 @@
+export type RadioButtonType<T, U> = {
+  title: U
+  value: T
+}
