--- src/types/components/radioButtonList/RadioButtonList.types.ts
@@ -1,4 +1,0 @@
-export type RadioButtonType = {
-  title: string
-  value: number
-}
